+++ requirements.txt
@@ -5,3 +5,4 @@ asyncpg>=0.29.0
 asyncpg-stubs>=0.29.1
 uvicorn>=0.24.0
 wavelink>=3.0.0
+aiohttp>=3.7.4,<4
