+++ routes/__init__.py
@@ -1,5 +1,4 @@
 from .eventsub import EventSub as EventSubView
+from .mbti import Mbti as MbtiView
 from .oauth import OAuth as OAuthView
 from .quotes import Quotes as QuotesView
-from .test import Test as TestView
-# from .mbti import Mbti as MbtiView (for this folder, import the class as a view)
